// bao-source/bao-core/src/generated/flatbuffers/baohaus/realtime/v1/position3-d.ts
// automatically generated by the FlatBuffers compiler, do not modify
import { Builder } from '@baohaus/flatbuf-bao/builder';
import type { IGeneratedObject, Offset } from '@baohaus/flatbuf-bao/builder';
import { ByteBuffer } from '@baohaus/flatbuf-bao/reader';
import type { IUnpackableObject } from '@baohaus/flatbuf-bao/reader';
export class Position3D implements IUnpackableObject<Position3DT> {
  bb: ByteBuffer | null = null;
  bb_pos = 0;
  __init(i: number, bb: ByteBuffer): Position3D {
    this.bb_pos = i;
    this.bb = bb;
    return this;
  }

  latitude(): number {
    return this.bb!.readFloat64(this.bb_pos);
  }

  longitude(): number {
    return this.bb!.readFloat64(this.bb_pos + 8);
  }

  altitude(): number {
    return this.bb!.readFloat64(this.bb_pos + 16);
  }

  relativeAltitude(): number {
    return this.bb!.readFloat64(this.bb_pos + 24);
  }

  static sizeOf(): number {
    return 32;
  }

  static createPosition3D(
    builder: Builder,
    latitude: number,
    longitude: number,
    altitude: number,
    relativeAltitude: number,
  ): Offset {
    builder.prep(8, 32);
    builder.writeFloat64(relativeAltitude);
    builder.writeFloat64(altitude);
    builder.writeFloat64(longitude);
    builder.writeFloat64(latitude);
    return builder.offset();
  }

  unpack(): Position3DT {
    return new Position3DT(
      this.latitude(),
      this.longitude(),
      this.altitude(),
      this.relativeAltitude(),
    );
  }

  unpackTo(_o: Position3DT): void {
    _o.latitude = this.latitude();
    _o.longitude = this.longitude();
    _o.altitude = this.altitude();
    _o.relativeAltitude = this.relativeAltitude();
  }
}

export class Position3DT implements IGeneratedObject {
  constructor(
    public latitude: number = 0.0,
    public longitude: number = 0.0,
    public altitude: number = 0.0,
    public relativeAltitude: number = 0.0,
  ) {}

  pack(builder: Builder): Offset {
    return Position3D.createPosition3D(
      builder,
      this.latitude,
      this.longitude,
      this.altitude,
      this.relativeAltitude,
    );
  }
}