// bao-source/bao-core/src/generated/flatbuffers/baohaus/realtime/v1/velocity3-d.ts
// automatically generated by the FlatBuffers compiler, do not modify
import { Builder } from '@baohaus/flatbuf-bao/builder';
import type { IGeneratedObject, Offset } from '@baohaus/flatbuf-bao/builder';
import { ByteBuffer } from '@baohaus/flatbuf-bao/reader';
import type { IUnpackableObject } from '@baohaus/flatbuf-bao/reader';
export class Velocity3D implements IUnpackableObject<Velocity3DT> {
  bb: ByteBuffer | null = null;
  bb_pos = 0;
  __init(i: number, bb: ByteBuffer): Velocity3D {
    this.bb_pos = i;
    this.bb = bb;
    return this;
  }

  x(): number {
    return this.bb!.readFloat32(this.bb_pos);
  }

  y(): number {
    return this.bb!.readFloat32(this.bb_pos + 4);
  }

  z(): number {
    return this.bb!.readFloat32(this.bb_pos + 8);
  }

  groundSpeed(): number {
    return this.bb!.readFloat32(this.bb_pos + 12);
  }

  airSpeed(): number {
    return this.bb!.readFloat32(this.bb_pos + 16);
  }

  static sizeOf(): number {
    return 20;
  }

  static createVelocity3D(
    builder: Builder,
    x: number,
    y: number,
    z: number,
    groundSpeed: number,
    airSpeed: number,
  ): Offset {
    builder.prep(4, 20);
    builder.writeFloat32(airSpeed);
    builder.writeFloat32(groundSpeed);
    builder.writeFloat32(z);
    builder.writeFloat32(y);
    builder.writeFloat32(x);
    return builder.offset();
  }

  unpack(): Velocity3DT {
    return new Velocity3DT(this.x(), this.y(), this.z(), this.groundSpeed(), this.airSpeed());
  }

  unpackTo(_o: Velocity3DT): void {
    _o.x = this.x();
    _o.y = this.y();
    _o.z = this.z();
    _o.groundSpeed = this.groundSpeed();
    _o.airSpeed = this.airSpeed();
  }
}

export class Velocity3DT implements IGeneratedObject {
  constructor(
    public x: number = 0.0,
    public y: number = 0.0,
    public z: number = 0.0,
    public groundSpeed: number = 0.0,
    public airSpeed: number = 0.0,
  ) {}

  pack(builder: Builder): Offset {
    return Velocity3D.createVelocity3D(
      builder,
      this.x,
      this.y,
      this.z,
      this.groundSpeed,
      this.airSpeed,
    );
  }
}